move tests from different components to settings
