+++ frontend_v2/src/app/components/challenge/challenge.component.spec.ts
@@ -20,17 +20,16 @@ import { ForceloginComponent } from '../../components/utility/forcelogin/forcelo
 import { FooterComponent } from '../../components/nav/footer/footer.component';
 import { TeamlistComponent } from '../../components/publiclists/teamlist/teamlist.component';
 import { SelectphaseComponent } from '../../components/utility/selectphase/selectphase.component';
-import { InputComponent } from '../../components/utility/input/input.component';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
-import { Observable, of } from 'rxjs';
+import { ActivatedRoute } from '@angular/router';
+import { of } from 'rxjs';
 import { MatTableModule } from '@angular/material';
 import { FormsModule } from '@angular/forms';
 
 describe('ChallengeComponent', () => {
   let component: ChallengeComponent;
   let fixture: ComponentFixture<ChallengeComponent>;
-  let router, authService, authServiceSpy;
+  let authService, authServiceSpy;
   const fakeActivatedRoute = {
     id: 1,
   };
@@ -76,20 +75,12 @@ describe('ChallengeComponent', () => {
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ChallengeComponent);
-    router = TestBed.get(Router);
     challengeService = TestBed.get(ChallengeService);
     globalService = TestBed.get(GlobalService);
     authService = TestBed.get(AuthService);
     apiService = TestBed.get(ApiService);
     endpointsService = TestBed.get(EndpointsService);
     component = fixture.componentInstance;
-
-    spyOn(router, 'navigate');
-    spyOn(globalService, 'handleApiError');
-    spyOn(globalService, 'showToast');
-    spyOn(globalService, 'showModal');
-    spyOn(globalService, 'showConfirm');
-    spyOn(endpointsService, 'editChallengeDetailsURL');
   });
 
   it('should create', () => {
@@ -99,10 +90,6 @@ describe('ChallengeComponent', () => {
   it('Global variables', () => {
     expect(component.isStarred).toBeFalsy();
     expect(component.isChallengeHost).toBeFalsy();
-    expect(component.publishChallenge).toEqual({
-      state: 'Not Published',
-      icon: 'fa fa-eye-slash red-text',
-    });
     expect(component.isParticipated).toBeFalsy();
     expect(component.isLoggedIn).toBeFalsy();
   });
@@ -135,166 +122,4 @@ describe('ChallengeComponent', () => {
     component.starToggle(challengeId);
     expect(globalService.showToast).toHaveBeenCalledWith('error', 'Please login to star the challenge!', 5);
   });
-
-  it('should toggle the publish challenge state from public to private', async(() => {
-    component.publishChallenge = {
-      state: 'Published',
-      icon: 'fa fa-eye green-text',
-    };
-    const expectedSuccessMsg = 'The challenge was successfully made private';
-    spyOn(apiService, 'patchUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.next({ results: [{}] });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.togglePublishChallengeState();
-    fixture.detectChanges();
-    expect(globalService.showConfirm).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.patchUrl).toHaveBeenCalled();
-    expect(component.publishChallenge.state).toEqual('Not Published');
-    expect(component.publishChallenge.icon).toEqual('fa fa-eye-slash red-text');
-    expect(globalService.showToast).toHaveBeenCalledWith('success', expectedSuccessMsg, 5);
-  }));
-
-  it('should toggle the publish challenge state from private to public', async(() => {
-    component.publishChallenge = {
-      state: 'Not Published',
-      icon: 'fa fa-eye-slash red-text',
-    };
-    const expectedSuccessMsg = 'The challenge was successfully made public';
-    spyOn(apiService, 'patchUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.next({ results: [{}] });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.togglePublishChallengeState();
-    fixture.detectChanges();
-    expect(globalService.showConfirm).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.patchUrl).toHaveBeenCalled();
-    expect(component.publishChallenge.state).toEqual('Published');
-    expect(component.publishChallenge.icon).toEqual('fa fa-eye green-text');
-    expect(globalService.showToast).toHaveBeenCalledWith('success', expectedSuccessMsg, 5);
-  }));
-
-  it('should handle the API error for `togglePublishChallengeState` method', async(() => {
-    const expectedApiError = {
-      error: 'Api error',
-    };
-    spyOn(apiService, 'patchUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.error({ error: expectedApiError.error });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.togglePublishChallengeState();
-    fixture.detectChanges();
-    expect(globalService.showConfirm).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.patchUrl).toHaveBeenCalled();
-    expect(globalService.handleApiError).toHaveBeenCalledWith(expectedApiError, true);
-    expect(globalService.showToast).toHaveBeenCalledWith('error', expectedApiError);
-  }));
-
-  it('should show the modal and successfully edit the title of a challenge', async(() => {
-    component.challenge = {
-      id: 1,
-      title: 'Challenge title',
-      creator: 'Host user',
-      description: 'Challenge description',
-    };
-    const expectedUpdateTitle = 'Updated challenge title';
-    const expectedSuccessMsg = 'The challenge title is  successfully updated!';
-    spyOn(apiService, 'patchUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.next({ title: expectedUpdateTitle });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.editChallengeTitle();
-    fixture.detectChanges();
-    expect(globalService.showModal).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.patchUrl).toHaveBeenCalled();
-    expect(component.challenge.title).toEqual(expectedUpdateTitle);
-    expect(globalService.showToast).toHaveBeenCalledWith('success', expectedSuccessMsg, 5);
-  }));
-
-  it('should handle the API error for `editChallengeTitle` method', async(() => {
-    component.challenge = {
-      id: 1,
-      title: 'Challenge title',
-      creator: 'Host user',
-      description: 'Challenge description',
-    };
-    const expectedApiError = {
-      error: 'Api error',
-    };
-    spyOn(apiService, 'patchUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.error({ error: expectedApiError.error });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.editChallengeTitle();
-    fixture.detectChanges();
-    expect(globalService.showModal).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.patchUrl).toHaveBeenCalled();
-    expect(globalService.handleApiError).toHaveBeenCalledWith(expectedApiError, true);
-    expect(globalService.showToast).toHaveBeenCalledWith('error', expectedApiError);
-  }));
-
-  it('should show the modal with the form fields and successfully delete the challenge', async(() => {
-    const expectedSuccessMsg = 'The Challenge is successfully deleted!';
-    spyOn(apiService, 'postUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.next({ results: [{}] });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.deleteChallenge();
-    fixture.detectChanges();
-    expect(globalService.showModal).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.postUrl).toHaveBeenCalled();
-    expect(router.navigate).toHaveBeenCalledWith(['/challenges/all']);
-    expect(globalService.showToast).toHaveBeenCalledWith('success', expectedSuccessMsg, 5);
-  }));
-
-  it('should handle the API error for `deleteChallenge` method', async(() => {
-    const expectedApiError = {
-      error: 'Api error',
-    };
-    spyOn(apiService, 'postUrl').and.returnValue(
-      new Observable((observer) => {
-        observer.error({ error: expectedApiError.error });
-        observer.complete();
-        return { unsubscribe() {} };
-      })
-    );
-
-    component.deleteChallenge();
-    fixture.detectChanges();
-    expect(globalService.showModal).toHaveBeenCalled();
-    component.apiCall();
-    expect(apiService.postUrl).toHaveBeenCalled();
-    expect(globalService.handleApiError).toHaveBeenCalledWith(expectedApiError, true);
-    expect(globalService.showToast).toHaveBeenCalledWith('error', expectedApiError);
-  }));
 });
